feat: move to new page, clean up navbar for mobile
