+++ src/pages/profile.tsx
@@ -3,7 +3,6 @@ import { type Book } from "../types";
 import type { ProfileViewDetailed } from "@atproto/api/dist/client/types/app/bsky/actor/defs";
 import { BookList } from "./components/book";
 import { ProfileHeader } from "./components/ProfileHeader";
-import { GoodreadsImport } from "./components/GoodreadsImport";
 import { BookReview } from "./components/BookReview";
 
 export const ProfilePage: FC<{
@@ -11,15 +10,12 @@ export const ProfilePage: FC<{
   books: Book[];
   isBuzzer: boolean;
   profile: ProfileViewDetailed | null;
-  isOwner: boolean;
-}> = ({ handle, profile, books, isBuzzer, isOwner }) => {
+}> = ({ handle, profile, books, isBuzzer }) => {
   return (
     <div class="bg-sand container mx-auto min-h-[calc(100vh-64px)] max-w-7xl p-8 px-3 dark:bg-zinc-900 dark:text-white">
       <ProfileHeader handle={handle} profile={profile} books={books} />
 
       <div class="flex flex-col gap-10">
-        {isOwner && <GoodreadsImport />}
-
         {isBuzzer ? (
           <>
             <section class="mt-8 flex flex-col gap-2 px-4 lg:px-8">
