+++ src/routes.tsx
@@ -36,6 +36,7 @@ import { type HiveId } from "./types";
 import { updateBookRecord } from "./utils/getBook";
 import { getProfile } from "./utils/getProfile";
 import { readThroughCache } from "./utils/readThroughCache";
+import { GoodreadsImport } from "./pages/import";
 
 declare module "hono" {
   interface ContextRenderer {
@@ -425,14 +426,34 @@ export function createRouter(app: HonoServer) {
     return c.redirect(`/profile/${handle}`);
   });
 
+  // Redirect to profile/:handle
+  app.get("/import", async (c) => {
+    const agent = await c.get("ctx").getSessionAgent();
+    if (!agent) {
+      return c.html(
+        <Layout>
+          <ErrorPage
+            message="Invalid Session"
+            description="Login to view your profile"
+            statusCode={401}
+          />
+        </Layout>,
+        401,
+      );
+    }
+
+    return c.render(<GoodreadsImport />, {
+      title: "BookHive | Import",
+      description: "Import your Goodreads library to BookHive",
+    });
+  });
+
   app.get("/profile/:handle", async (c) => {
     const handle = c.req.param("handle");
 
     const did = isDid(handle)
       ? handle
       : await c.get("ctx").baseIdResolver.handle.resolve(handle);
-
-    const agent = await c.get("ctx").getSessionAgent();
 
     if (!did) {
       return c.render(
@@ -456,17 +477,6 @@ export function createRouter(app: HonoServer) {
         .executeTakeFirst(),
     );
 
-    // if (!profileRecord) {
-    //   return c.render(
-    //     <Fragment>
-    //       <h1>Profile {handle} not found</h1>
-    //       <p>
-    //         This profile may not exist or has not logged any books on bookhive
-    //       </p>
-    //     </Fragment>,
-    //   );
-    // }
-
     const profile = await getProfile({ ctx: c.get("ctx"), did });
     const books = isBuzzer
       ? await c
@@ -486,7 +496,6 @@ export function createRouter(app: HonoServer) {
         handle={handle}
         books={books}
         profile={profile}
-        isOwner={agent?.did === did}
       />,
       {
         title: "BookHive | @" + handle,
